feat: Add join route for community membership
This commit adds a new route `/join/:id` to allow users to join a community by providing the community ID. It checks if the user is authenticated and returns an error if not. It also checks if the user is already a member of the community and returns an error if so. If the user is not a member, it inserts a new community member record with the user ID, community ID, and role set to "member".