feat: Add join route for trip membership
This commit adds a new route `/trip/join/:id` to allow users to join a trip by providing the trip ID. It checks if the user is authenticated and returns an error if not. It also checks if the user is already a member of the trip and returns an error if so. If the user is not a member, it inserts a new trip participant record with the user ID, trip ID, and role set to "participant".