feat: Add user vehicle creation route
This commit adds a new route for creating user vehicles. The route is accessible via a POST request to "/user/vehicle" and requires authentication. It allows users to add vehicles to their profiles by providing the vehicle ID and year. This feature enhancement improves the functionality of the application by enabling users to associate vehicles with their profiles.