feat: Add GET route for user vehicles
This commit adds a new GET route for retrieving user vehicles. The route is accessible via a GET request to "/user/vehicle" and requires authentication. It returns a list of vehicles associated with the user's profile. This feature enhancement improves the functionality of the application by allowing users to view their associated vehicles.