feat: Add userVehicles to userRoutes
This commit adds the `userVehicles` property to the `userRoutes` object in the `user.ts` file. The `userVehicles` property includes a nested `with` object that fetches additional data about the vehicles associated with the user's profile. This feature enhancement improves the functionality of the application by allowing users to retrieve and view their associated vehicles.