feat: Add Split type for splitting strings by delimiter
This commit adds the Split type to the types.ts file. The Split type is used to split a string into an array of substrings based on a delimiter. This type enhances the functionality of the application by providing a convenient way to split strings and process the resulting substrings separately.